% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Abalone}
\alias{Abalone}
\title{Abalone measurements}
\format{
A \code{data.frame} with 4177 rows and 9 columns:
\describe{
\item{\code{Sex}}{A factor with 3 levels: "M" (male), "F" (female) and "I" (infant).}
\item{\code{Length}}{Longest shell measurement (mm).}
\item{\code{Diameter}}{Perpendicular to length (mm).}
\item{\code{Height}}{With meat in shell (mm).}
\item{\code{Whole_weight}}{Whole abalone (g).}
\item{\code{Shucked_weight}}{Weight of meat (g).}
\item{\code{Viscera_weight}}{Gut weight (after bleeding, g).}
\item{\code{Shell_weight}}{After being dried (g).}
\item{\code{Rings}}{Number of rings. +1.5 gives the age in years.}
}
}
\source{
Available in UC Irvine Machine Learning Repository as
\href{https://archive.ics.uci.edu/dataset/1/}{dataset with ID 1}
}
\usage{
Abalone
}
\description{
Predict the age of abalone from physical measurements.
}
\details{
The age of abalone is determined by cutting the shell through the cone,
staining it, and counting the number of rings through a microscope -
a boring and time-consuming task.  Other measurements, which are easier to
obtain, are used to predict the age.  Further information, such as weather
patterns and location (hence food availability) may be required to solve the
problem. From the original data examples with missing values were removed
(the majority having the predicted value missing), and the ranges of the
continuous values have been scaled for use with an ANN (by dividing by 200).
}
\keyword{datasets}
