% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absences}
\alias{absences}
\title{Student absences}
\format{
A \code{data.frame} with 8 rows and 2 columns:
\describe{
\item{\code{n_absences}}{Number of absences.}
\item{\code{n_students}}{Number of students with a given absence level.}
}
}
\usage{
absences
}
\description{
Absences of students in a semester.
}
\keyword{datasets}
