% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DebTrivedi}
\alias{DebTrivedi}
\title{Medical care demand by the elderly}
\format{
A \code{data.frame} with 4406 rows and 19 columns:
\describe{
\item{\code{ofp}}{Number of physician office visits.}
\item{\code{ofnp}}{Number of non-physician office visits.}
\item{\code{opp}}{Number of physician hospital outpatient visits.}
\item{\code{opnp}}{Number of non-physician hospital outpatient visits.}
\item{\code{emr}}{Number of emergency room visits.}
\item{\code{hosp}}{Number of hospital stays.}
\item{\code{health}}{Self-perceived health status. A factor with 3 levels: "poor",
"average" and "excellent".}
\item{\code{numchron}}{Number of chronic conditions (cancer, heart attack, gall bladder problems,
emphysema, arthritis, diabetes, other heart disease).}
\item{\code{adldiff}}{Does the person have a condition that limits activities of daily living?
A factor with 2 levels "Yes" and "No".}
\item{\code{region}}{Region of residence. A factor with 4 levels: "noreast", "midwest", "south" and "west".}
\item{\code{age}}{Age in years (divided by 10).}
\item{\code{black}}{Is the person black? A factor with 2 levels "Yes" and "No".}
\item{\code{male}}{Is the person male? A factor with 2 levels "Yes" and "No".}
\item{\code{married}}{Is the person married? A factor with 2 levels "Yes" and "No".}
\item{\code{school}}{Number of years of education.}
\item{\code{faminc}}{Family income in $10,000.}
\item{\code{employed}}{Is the person employed? A factor with 2 levels "Yes" and "No".}
\item{\code{privins}}{Is the person covered by private health insurance?
A factor with 2 levels "Yes" and "No".}
\item{\code{medicaid}}{Is the person covered by Medicaid? A factor with 2 levels "Yes" and "No".}
}
}
\source{
Partha Deb and Pravin K. Trivedi,
\emph{Demand for Medical Care by the Elderly: A Finite Mixture Approach},
Journal of Applied Econometrics, Vol. 12, No. 3, 1997, pp. 313-336.
\url{http://qed.econ.queensu.ca/jae/1997-v12.3/deb-trivedi/}
}
\usage{
DebTrivedi
}
\description{
The dataset consists of a subsample of elderly individuals
(66 and older) drawn from the National Medical Expenditure Survey (1987).
Dummy variable columns from the original dataset have been converted
to appropriate factors.
}
\keyword{datasets}
