% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goals}
\alias{goals}
\title{Goals from 64 games}
\format{
A numeric vector of length 64.
}
\usage{
goals
}
\description{
Goals won by a football team in 64 subsequent games (simulated data).
}
\keyword{datasets}
