% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goals2}
\alias{goals2}
\title{Goals from 200 games}
\format{
A numeric vector of length 200.
}
\usage{
goals2
}
\description{
Goals won by a football team in 200 subsequent games (simulated data).
}
\keyword{datasets}
