% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{hdi.distribution}
\alias{hdi.distribution}
\title{Compute the HDI for a distribution}
\usage{
hdi.distribution(d, credMass = 0.95, ...)
}
\arguments{
\item{d}{A distribution from \code{distributions3} package.}

\item{credMass}{A scalar from \verb{[0,1]} specifying the mass within the credible
interval. Defaults to 0.95.}

\item{...}{Other arguments passed to \code{\link[HDInterval:hdi]{HDInterval::hdi()}}.}
}
\description{
Compute the HDI for a distribution
}
\seealso{
\code{\link[HDInterval:hdi]{HDInterval::hdi()}}
}
