% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HtWtData300}
\alias{HtWtData300}
\title{Height/weigth data - 300 people}
\format{
An object of class \code{data.frame} with 300 rows and 3 columns.
}
\usage{
HtWtData300
}
\description{
Height/weigth data - 300 people
}
\keyword{datasets}
