% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBAPlayers}
\alias{NBAPlayers}
\title{NBA players 2022/23}
\format{
A tibble (a data frame) with 539 rows and 31 columns.
}
\source{
\url{https://www.basketball-reference.com/leagues/NBA_2023_per_game.html}
}
\usage{
NBAPlayers
}
\description{
NBA player per-game statistics from 2022/23 regular season.
}
\details{
For players that changed teams during the regular season only total
scores are shown (for such players \code{Tm == TOT}). Some players were fielded
at two different positions: then only the first position is shown.
}
\keyword{datasets}
