% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Obesity}
\alias{Obesity}
\title{Obesity levels}
\format{
A \code{data.frame} with 2111 rows and 17 columns:
\describe{
\item{\code{Gender}}{Sex. A factor with 2 levels: "Female" and "Male".}
\item{\code{Age}}{Age in years.}
\item{\code{Height}}{Height in meters.}
\item{\code{Weight}}{Weight in kgs.}
\item{\code{family_history_with_overweight}}{Has a family member suffered or
does he/she suffer from overweight? A factor with 2 levels: "yes" and "no".}
\item{\code{FAVC}}{Does the person eat high caloric food frequently?
A factor with 2 levels: "yes" and "no".}
\item{\code{FCVC}}{Does the person usually eat vegetables in your meals?
1 - never, 2 - sometimes, 3 - always.}
\item{\code{NCP}}{How many main meals does the person have daily?
A positive integer.}
\item{\code{CAEC}}{Does the person eat any food between meals? A factor with
4 levels: "no", "Sometimes", "Frequently" and "Always".}
\item{\code{SMOKE}}{Does the person smoke?  A factor with 2 levels: "yes" and "no".}
\item{\code{CH2O}}{How much water does the person drink daily?
1 - less than a liter, 2 - between 1 and 2 liters, 3 - more than 3 liters.}
\item{\code{SCC}}{Does the person monitor the calories he/she eats daily?
A factor with 2 levels: "yes" and "no".}
\item{\code{FAF}}{How often does the person have physical activity?
0 - never, 1 - 1 or 2 days, 2 - 2 or 4 days, 3 - 4 or 5 days.}
\item{\code{TUE}}{How much time does the person use technological devices such as
cell phone, videogames, television, computer and others?
0 - 0-2 hours, 1 - 3-5 hours, 2 - more than 5 hours.}
\item{\code{CALC}}{How often does the person drink alcohol? A factor with
4 levels: "no", "Sometimes", "Frequently", "Always".}
\item{\code{MTRANS}}{Which transportation does the person usually use?
A factor with 5 levels: "Automobile", "Bike", "Motorbike",
"Public_Transportation", "Walking".}
\item{\code{NObeyesdad}}{Obesity level. An ordered factor with levels:
"Insufficient_Weight" < "Normal_Weight" < "Overweight_Level_I" <
"Overweight_Level_II" < "Obesity_Type_I" < "Obesity_Type_II" <
"Obesity_Type_III"}
}
}
\source{
Fabio Mendoza Palechor and Alexis de la Hoz Manotas,
\emph{Dataset for estimation of obesity levels based on eating habits and physical
condition in individuals from Colombia, Peru and Mexico}, Data in Brief,
Vol. 25, 2019, pp. 104344.
\href{https://doi.org/10.1016/j.dib.2019.104344}{DOI}

Available in
UC Irvine Machine Learning Repository as
\href{https://archive.ics.uci.edu/dataset/544/}{dataset with ID 544}
}
\usage{
Obesity
}
\description{
Estimation of obesity levels based on eating habits and physical condition.
}
\details{
This set contains data for the estimation of obesity levels in individuals
from the countries of Mexico, Peru and Colombia, based on their eating habits
and physical condition.
77\% of the data was generated synthetically using the Weka tool and the
SMOTE filter, 23\% of the data was collected directly from users through a
web platform. For discussion and more information of
the dataset creation, please refer to the full-length article cited in the
Source section (see below).
The original UCI ML dataset has been coerced to a regular R data frame by
turning categorical and binary variables into appropriate factors.
}
\keyword{datasets}
