% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plates}
\alias{plates}
\title{Plate thickness}
\format{
A numeric vector of length 25.
}
\usage{
plates
}
\description{
Thicknesses of metal plates coming from a production line, in millimeters
(simulated data).
}
\keyword{datasets}
