% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PremierLeague2006}
\alias{PremierLeague2006}
\title{Premier League 2006/7}
\format{
A tibble with 380 rows and 24 columns.
}
\source{
\url{https://www.kaggle.com/datasets/saife245/english-premier-league/}
}
\usage{
PremierLeague2006
}
\description{
English Premier League game statistics from 2006/7 season.
}
\seealso{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\keyword{datasets}
