% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PremierLeague2021}
\alias{PremierLeague2021}
\title{Premier League 2021/22}
\format{
A tibble with 380 rows and 107 columns.
}
\source{
\url{https://www.kaggle.com/datasets/saife245/english-premier-league/}
}
\usage{
PremierLeague2021
}
\description{
English Premier League game statistics from 2021/22 season.
}
\seealso{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\keyword{datasets}
