% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_utils.R
\name{show_central}
\alias{show_central}
\title{Adds central value indicators to the current plot}
\usage{
show_central(value, y, type = "mean", digits = 3, col = "blue")
}
\arguments{
\item{value}{Numeric. The central value.}

\item{y}{Numeric. \code{y} coordinate of the text.}

\item{type}{Character. Type of central value: "mean" or "mode".}

\item{digits}{Integer. Number of significant digits to show. Defaults to 3.}

\item{col}{The color to be used. Defaults to "blue".}
}
\description{
Adds central value indicators to the current plot
}
