% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_utils.R
\name{show_hdi}
\alias{show_hdi}
\title{Adds HDI to the current plot}
\usage{
show_hdi(hdi, y = 0, digits = 3, pos = 3, col = "red")
}
\arguments{
\item{hdi}{Numeric of length 2. The Highest Density Interval.}

\item{y}{Numeric. \code{y} coordinate of the line.}

\item{digits}{Integer. Number of significant digits to show. Defaults to 3.}

\item{pos}{A position specifier for the text (see \code{\link[graphics:text]{graphics::text()}}). Defaults to 3.}

\item{col}{The color to be used. Defaults to "red".}
}
\description{
Adds HDI to the current plot
}
