% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_utils.R
\name{show_mle}
\alias{show_mle}
\title{Adds a point estimator (e.g. MLE) to the current plot}
\usage{
show_mle(mle, y = 0, digits = 3, pos = 3, col = "orange")
}
\arguments{
\item{mle}{Numeric scalar. The point estimator.}

\item{y}{Numeric. \code{y} coordinate of the text.}

\item{digits}{Integer. Number of significant digits to show. Defaults to 3.}

\item{pos}{A position specifier for the text (see \code{\link[graphics:text]{graphics::text()}}). Defaults to 3.}

\item{col}{The color to be used. Defaults to "orange".}
}
\description{
Adds a point estimator (e.g. MLE) to the current plot
}
