% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TherapeuticTouch}
\alias{TherapeuticTouch}
\title{Data from an experiment on the therapeutic touch}
\format{
An object of class \code{data.frame} with 280 rows and 2 columns.
}
\usage{
TherapeuticTouch
}
\description{
Data from an experiment on the therapeutic touch
}
\keyword{datasets}
