% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_coins}
\alias{two_coins}
\title{Two coins}
\format{
A \code{data.frame} with 15 rows and 2 columns:
\describe{
\item{\code{y}}{Toss result: 1 for Heads, 0 for Tails.}
\item{\code{coin}}{Coin ID. A factor with 2 levels.}
}
}
\source{
John K. Kruschke,
\emph{Doing Bayesian Data Analysis: A Tutorial with R, JAGS  and Stan (2nd Edition)},
Academic Press, 2014.
\url{https://sites.google.com/site/doingbayesiandataanalysis/software-installation}
}
\usage{
two_coins
}
\description{
Tosses of two coins with different load.
}
\keyword{datasets}
