% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WineQuality}
\alias{WineQuality}
\title{Wine quality}
\format{
A \code{data.frame} with 6497 rows and 13 columns:
\describe{
\item{\code{fixed_acidity}}{Amount (g) of tartaric acid in dm^3.}
\item{\code{volatile_acidity}}{Amount (g) of acetic acid in dm^3.}
\item{\code{citric_acid}}{g/dm^3.}
\item{\code{residual_sugar}}{g/dm^3.}
\item{\code{chlorides}}{Amount (g) of sodium chloride in dm^3.}
\item{\code{free_sulfur_dioxide}}{mg/dm^3.}
\item{\code{total_sulfur_dioxide}}{mg/dm^3.}
\item{\code{density}}{g/cm^3.}
\item{\code{pH}}{Measure of acidity or alkalinity.}
\item{\code{sulphates}}{Amount (g) of potassium sulphate in dm^3.}
\item{\code{alcohol}}{Volume percentage.}
\item{\code{quality}}{Sensory score between 0 and 10 (actually between 3 and 9) -
median of at least 3 blind evaluations. An ordered factor.}
\item{\code{color}}{A factor with 2 levels: "red" and "white".}
}
}
\source{
P. Cortez, Antonio Luiz Cerdeira, Fernando Almeida, Telmo Matos and Jose Reis,
\emph{Modeling wine preferences by data mining from physicochemical properties},
Decision Support Systems, Vol. 47, 2009, pp. 547-553.
\href{https://doi.org/10.1016/j.dss.2009.05.016}{DOI}

Available in UC Irvine Machine Learning Repository as
\href{https://archive.ics.uci.edu/dataset/186/}{dataset with ID 186}
}
\usage{
WineQuality
}
\description{
This dataset contains measurements of samples of \emph{vinho verde}, a unique
wine from the Minho (north-west) region of Portugal.
The data were collected from May 2004 to February 2007 using only protected
designation of origin samples that were tested at the oﬃcial certification
entity (CVRVV).
}
\details{
All numerical columns (\code{fixed_acidity} to \code{alcohol}) have been standardized
to facilitate the use of ordinal regression and other ML models.
Nevertheless, their
description refers to the original measurements.
}
\keyword{datasets}
