% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{absences}
\alias{absences}
\title{Absences of students}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\usage{
absences
}
\description{
Absences of students
}
\keyword{datasets}
