% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goals2}
\alias{goals2}
\title{Goals won by a football team in 200 subsequent games}
\format{
An object of class \code{numeric} of length 200.
}
\usage{
goals2
}
\description{
Goals won by a football team in 200 subsequent games
}
\keyword{datasets}
