% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HtWtData30}
\alias{HtWtData30}
\title{Height/weigth data - 30 people}
\format{
An object of class \code{data.frame} with 30 rows and 3 columns.
}
\usage{
HtWtData30
}
\description{
Height/weigth data - 30 people
}
\keyword{datasets}
