% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBAPlayers}
\alias{NBAPlayers}
\title{NBA player statistics from 2022/23 regular season}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 539 rows and 31 columns.
}
\usage{
NBAPlayers
}
\description{
NBA player statistics from 2022/23 regular season
}
\keyword{datasets}
