% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plates}
\alias{plates}
\title{Thicknesses of metal plates in millimeters}
\format{
An object of class \code{numeric} of length 25.
}
\usage{
plates
}
\description{
Thicknesses of metal plates in millimeters
}
\keyword{datasets}
