% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PremierLeague2006}
\alias{PremierLeague2006}
\title{English Premier League game statistics from 2006/7 season}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 380 rows and 24 columns.
}
\usage{
PremierLeague2006
}
\description{
English Premier League game statistics from 2006/7 season
}
\keyword{datasets}
