% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PremierLeague2021}
\alias{PremierLeague2021}
\title{English Premier League game statistics from 2021/22 season}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 380 rows and 107 columns.
}
\usage{
PremierLeague2021
}
\description{
English Premier League game statistics from 2021/22 season
}
\keyword{datasets}
