% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_coins}
\alias{two_coins}
\title{Tosses of two different coins}
\format{
An object of class \code{data.frame} with 15 rows and 2 columns.
}
\usage{
two_coins
}
\description{
Tosses of two different coins
}
\keyword{datasets}
